@@ -74,27 +72,29 @@
     "cypress": "^9.1.0",
     "dotenv": "^16.0.0",
     "dotenv-webpack": "^7.1.0",
-    "eslint": "~8.7.0",
-    "eslint-config-next": "12.1.1-canary.3",
+    "eslint": "8.15.0",
+    "eslint-config-next": "12.3.1",
     "eslint-config-prettier": "^8.1.0",
     "eslint-config-prettier-standard": "^4.0.1",
     "eslint-config-standard": "^16.0.3",
     "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jsx-a11y": "6.5.1",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsx-a11y": "6.6.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-promise": "^5.2.0",
-    "eslint-plugin-react": "7.29.3",
-    "eslint-plugin-react-hooks": "4.3.0",
+    "eslint-plugin-react": "7.31.8",
+    "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-testing-library": "^5.10.0",
-    "jest": "29.4.1",
+    "jest": "28.1.3",
     "next-absolute-url": "^1.2.2",
-    "prettier": "2.5.1",
+    "prettier": "2.8.4",
     "prettier-config-standard": "^5.0.0",
-    "react-test-renderer": "17.0.2",
-    "ts-jest": "29.0.5",
-    "typescript": "~4.5.2"
+    "react-test-renderer": "18.2.0",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.4",
+    "nx": "14.8.4"
   },
   "dependencies": {
     "@radix-ui/colors": "^0.1.8",
@@ -105,16 +105,17 @@
     "govuk-frontend": "^4.4.1",
     "gray-matter": "^4.0.3",
     "isomorphic-unfetch": "^4.0.2",
-    "jest-environment-jsdom": "^29.4.1",
+    "jest-environment-jsdom": "28.1.3",
     "lint": "^0.7.0",
     "lodash": "^4.17.21",
     "marksy": "^8.0.0",
-    "next": "12.1.0",
+    "next": "12.3.1",
     "next-seo": "^5.15.0",
     "nextjs-url": "^1.0.4",
-    "react": "17.0.2",
+    "react": "18.2.0",
     "react-contentful": "^2.0.31",
-    "react-dom": "17.0.2",
+    "react-dom": "18.2.0",
+    "react-is": "18.0.0",
     "react-json-view-ssr": "^1.19.1",
     "react-markdown": "^8.0.4",
     "regenerator-runtime": "0.13.7",
