@@ -1,15 +1,16 @@
-module.exports = {
+/* eslint-disable */
+export default {
   displayName: 'environment-variables',
   preset: '../../jest.preset.js',
   globals: {
     'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json'
-    }
+      tsconfig: '<rootDir>/tsconfig.spec.json',
+    },
   },
   transform: {
     '^(?!.*\\.(js|jsx|ts|tsx|css|json)$)': '@nrwl/react/plugins/jest',
-    '^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nrwl/next/babel'] }]
+    '^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nrwl/next/babel'] }],
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'd.ts'],
-  coverageDirectory: '../../coverage/libs/environment-variables'
-}
+  coverageDirectory: '../../coverage/libs/environment-variables',
+};
